backend/lineart/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
onnxruntime==1.17.3
numpy==1.26.4
Pillow==10.3.0
